Add files via upload
The nested dictionary was successfully unraveled where the first elevation of the first model of the first unit was indexed as an example with the created variables for them.  This is a good starting point so that specific info like id's, codes and prices at the level of unit or model or elevation can be iterated and put into a Pandas DataFrame for analysis.